requirements.txt
anyio==4.8.0
asgiref==3.8.1
certifi==2024.12.14
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pillow==11.1.0
python-telegram-bot==21.10
pytz==2024.2
sniffio==1.3.1
sqlparse==0.5.3
telegram==0.0.1
typing_extensions==4.12.2
